# hcp-portal/Models/PrescriberProfile.cs
namespace HcpPortal.Models;

public class PrescriberProfile
{
    public string PrescriberId { get; set; } = string.Empty;
    public string Npi { get; set; } = string.Empty;
    public string FullName { get; set; } = string.Empty;
    public string Specialty { get; set; } = string.Empty;
    public string ClinicName { get; set; } = string.Empty;
}