# hcp-portal/openapi/hcp-portal-v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "HcpPortal API",
    "version": "v1"
  },
  "paths": {
    "/api/Prescribers/{prescriberId}/profile": {
      "get": {
        "tags": [
          "Prescribers"
        ],
        "parameters": [
          {
            "name": "prescriberId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriberProfile"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriberProfile"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriberProfile"
                }
              }
            }
          }
        }
      }
    },
    "/api/Prescribers/orders": {
      "post": {
        "tags": [
          "Prescribers"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MedicationOrderRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MedicationOrderRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MedicationOrderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MedicationOrderResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedicationOrderResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedicationOrderResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "MedicationOrderRequest": {
        "type": "object",
        "properties": {
          "patientId": {
            "type": "string",
            "nullable": true
          },
          "prescriberId": {
            "type": "string",
            "nullable": true
          },
          "medicationCode": {
            "type": "string",
            "nullable": true
          },
          "medicationName": {
            "type": "string",
            "nullable": true
          },
          "strength": {
            "type": "string",
            "nullable": true
          },
          "directions": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "refills": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "MedicationOrderResponse": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "clinicalReviewToken": {
            "type": "string",
            "nullable": true
          },
          "submittedAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "PrescriberProfile": {
        "type": "object",
        "properties": {
          "prescriberId": {
            "type": "string",
            "nullable": true
          },
          "npi": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "specialty": {
            "type": "string",
            "nullable": true
          },
          "clinicName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  },
  "tags": [
    {
      "name": "Prescribers"
    }
  ]
}